Added Features!
First, I made it select stamps when you create them.

Second I added an option to add new colors. I added a new div for the new button and all of the color buttons.

Third, I made it auto select colors when they are clicked, so you don't have to modify them to select them.

Fourth, it auto clicks the pencil button when the page starts so it highlights the button and it makes more sense to default to drawing instead of stamps.

Lastly, I changed my new color adder to suit the current button structure. I don't know if this was the best way to do it since I needed to mod the loop that adds each of the buttons.